requirements.txt
asgiref==3.9.0
astroid==3.1.0
bcrypt==4.1.2
black==24.4.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.9.2
cryptography==42.0.8
dill==0.4.0
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.8
psycopg2-binary==2.9.9
pycodestyle==2.12.1
PyJWT==2.10.1
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.9.0
pytest==8.2.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0
